requirements.txt
requests
beautifulsoup4
moviepy
Pillow
